src/pyusd_analytics.py
"""
PYUSD Analytics Module

This module provides real-time analytics for PYUSD transactions using
Google BigQuery's public crypto datasets and real-time event tracking.
"""

from google.cloud import bigquery
from typing import Dict, List, Optional
import pandas as pd
import asyncio
from datetime import datetime, timedelta

class PYUSDAnalytics:
    def __init__(self):
        self.client = bigquery.Client()
        self.project = 'bigquery-public-data'
        self.dataset = 'crypto_ethereum'

    async def get_pyusd_transfers(self, 
                                days_back: int = 7) -> pd.DataFrame:
        """
        Query recent PYUSD transfers from BigQuery
        """
        query = f"""
        SELECT
            block_timestamp,
            from_address,
            to_address,
            value / 1e6 as amount,
            transaction_hash,
            gas_price
        FROM `{self.project}.{self.dataset}.token_transfers`
        WHERE token_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
        AND block_timestamp >= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL {days_back} DAY)
        ORDER BY block_timestamp DESC
        """
        
        query_job = self.client.query(query)
        return query_job.to_dataframe()

    async def analyze_liquidity_pools(self) -> Dict:
        """
        Analyze PYUSD liquidity pools on major DEXs
        """
        query = """
        WITH pool_stats AS (
            SELECT
                pool_address,
                token0_symbol,
                token1_symbol,
                reserve0 / power(10, token0_decimals) as token0_reserve,
                reserve1 / power(10, token1_decimals) as token1_reserve,
                block_timestamp
            FROM `{self.project}.{self.dataset}.dex_pools`
            WHERE token0_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
            OR token1_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
            ORDER BY block_timestamp DESC
            LIMIT 1000
        )
        SELECT
            pool_address,
            token0_symbol,
            token1_symbol,
            token0_reserve,
            token1_reserve,
            block_timestamp
        FROM pool_stats
        """
        
        query_job = self.client.query(query)
        return query_job.to_dataframe()

    async def get_network_metrics(self) -> Dict:
        """
        Get network-wide metrics for PYUSD usage
        """
        query = """
        SELECT
            DATE(block_timestamp) as date,
            COUNT(DISTINCT from_address) as unique_senders,
            COUNT(DISTINCT to_address) as unique_receivers,
            COUNT(*) as total_transfers,
            SUM(value / 1e6) as total_volume
        FROM `{self.project}.{self.dataset}.token_transfers`
        WHERE token_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
        AND block_timestamp >= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 30 DAY)
        GROUP BY date
        ORDER BY date DESC
        """
        
        query_job = self.client.query(query)
        return query_job.to_dataframe()

    async def track_realtime_events(self) -> None:
        """
        Track real-time PYUSD events using BigQuery's real-time features
        """
        query = """
        SELECT *
        FROM `{self.project}.{self.dataset}.streaming_token_transfers`
        WHERE token_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
        """
        
        query_job = self.client.query(query)
        return query_job.to_dataframe()

    async def analyze_gas_usage(self) -> pd.DataFrame:
        """
        Analyze gas usage patterns for PYUSD transactions
        """
        query = """
        SELECT
            DATE(block_timestamp) as date,
            AVG(gas_price) as avg_gas_price,
            AVG(receipt_gas_used) as avg_gas_used,
            COUNT(*) as num_transactions
        FROM `{self.project}.{self.dataset}.transactions` t
        JOIN `{self.project}.{self.dataset}.token_transfers` tt
        ON t.hash = tt.transaction_hash
        WHERE tt.token_address = '0x6c3ea9036406c282D277Dc14762a4379D5084619'
        AND block_timestamp >= TIMESTAMP_SUB(CURRENT_TIMESTAMP(), INTERVAL 30 DAY)
        GROUP BY date
        ORDER BY date DESC
        """
        
        query_job = self.client.query(query)
        return query_job.to_dataframe()
